requirements.txt
ultralytics
pillow
torch
streamlit